Replace Get/Set/Del/ClearNV() with NoRegistry
The monolithic NoModule API should be reduced in a way or another.
This simple registry class supercedes the "NV" part of the NoModule
API.

Even though keys() is not as efficient as the old iterator based API,
doesn't expose the internal implementation details. Thus, it is
possible to even change the backend without breaking binary
compatibility.